+++ src/platform_impl/linux/x11/event_processor.rs
@@ -501,13 +501,14 @@ impl<T: 'static> EventProcessor<T> {
             ffi::Expose => {
                 let xev: &ffi::XExposeEvent = xev.as_ref();
 
-                let window = xev.window;
-                let window_id = mkwid(window);
+                // Multiple Expose events may be received for subareas of a window.
+                // We issue `RedrawRequested` only for the last event of such a series.
+                if xev.count == 0 {
+                    let window = xev.window;
+                    let window_id = mkwid(window);
 
-                callback(Event::WindowEvent {
-                    window_id,
-                    event: WindowEvent::RedrawRequested,
-                });
+                    callback(Event::RedrawRequested(window_id));
+                }
             }
 
             ffi::KeyPress | ffi::KeyRelease => {